refactor: Move license to internal
