+++ internal/license/license.go
