+++ internal/license/types.go
